fix virtual_memory computation on windows in refresh_pids
When refreshing a process' memory usage, we use `WorkingSetSize` for
`memory` and `PrivateUsage` for `virtual_memory`.

However, when calling `refresh_pids`, another codepath refreshes those
values, and use `VirtualBytes` for `virtual_memory`. This is not the
right field and is generally equal to 2TB, making this measurement
completely wrong.

The fix is to use the same field as when refreshing a single process,
`PrivatePageCount` (which afaict reading the documentation, should be
the same thing as `PrivateUsage`).